Feat: mock up event CUD api (#8)
