--- src/main/java/com/kyungbiseo/event/application/service/EventCommandService.java
@@ -1,7 +1,0 @@
-package com.kyungbiseo.event.application.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class EventCommandService {
-}
