+++ src/main/java/com/kyungbiseo/event/web/dto/EventUpdateRequest.java
@@ -1,0 +1,13 @@
+package com.kyungbiseo.event.web.dto;
+
+import java.time.LocalDateTime;
+
+import com.kyungbiseo.event.domain.EventPriority;
+import com.kyungbiseo.event.domain.EventType;
+
+public record EventUpdateRequest(Long friendId,
+								 String name,
+								 EventType type,
+								 LocalDateTime scheduledAt,
+								 EventPriority priority) {
+}
